Add Thai translation
There was a mistake when I was making branching before sending PRs. I was going to separated the translation and the art into separated branches and send a separated PRs (to reflect to a different issue), but I was merge the translation branch *before* branching into new art branch. Therefore, make the translation missing from the repository. I will mark this one as "invalid" because of duplication and it won't count toward Hacktoberfest contribution